requirements.txt
flask==3.1.2
flask-cors==4.0.0
python-dotenv==1.0.0
cloudinary==1.40.0
requests==2.31.0
apscheduler==3.10.4
werkzeug==3.1.5
gunicorn==22.0.0
